python-cli/requirements.txt
openai>=1.0.0
pyaudio>=0.2.11
keyboard>=0.13.5
pyautogui>=0.9.54
pynput>=1.7.6
python-dotenv>=1.0.0
tenacity>=8.2.0
requests>=2.31.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0